1. Modify: hp int to double 2. Add: Kinoko, Wand, Wizard class

+++ lib/hero/hero.dart
@@ -1,6 +1,6 @@
 class Hero {
   String name;
-  int hp;
+  double hp;
 
   Hero({required this.name, this.hp = 0});
 }
